+++ .gitignore
@@ -41,5 +41,3 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
-
-/assets/.env
