--- assets/.env.example
@@ -1,4 +1,0 @@
-URL= # URL/Link of Odoo
-DB= # Name of Database
-USER= # Username (usually email)
-PASS= # Password
